Refactor package structure and update data types in ledger components + user flow

+++ src/main/java/org/payment/settlement/Settlement.java
@@ -1,9 +1,9 @@
 package org.payment.settlement;
 
 import lombok.*;
-import org.payment.constant.Settlement.SettlementStatusEnum;
-
+import org.payment.consVar.settlement.SettlementStatusEnum;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.UUID;
 
@@ -18,7 +18,7 @@ public class Settlement {
     private UUID reference;
     private String fromActorId;
     private String toActorId;
-    private BigDecimal amount;
+    private BigInteger amount;
     private SettlementStatusEnum status; // PENDING, SETTLED, FAILED, ATTENTION_REQUIRED
     private Timestamp createdAt;
     private Timestamp updatedAt;
